solve to the practice for week 4

+++ src/main/java/com/practice/spring/util/validator/note/NoteLimitValidator.java
@@ -2,6 +2,6 @@ package com.practice.spring.util.validator.note;
 
 import com.practice.spring.util.validator.Validator;
 
-public interface NoteLimitValidator extends Validator<Integer> {
+public interface NoteLimitValidator extends Validator<Long> {
 
 }
